Factor bytes and errors into Common submodule
