+++ src/Common.jl
@@ -1,0 +1,11 @@
+"""
+Internal implementation detail.
+"""
+module Common
+
+using Compat
+
+include("bytes.jl")
+include("errors.jl")
+
+end
