+++ src/JSON.jl
@@ -6,9 +6,10 @@ using Compat
 
 export json # returns a compact (or indented) JSON representation as a string
 
+include("Common.jl")
 include("Parser.jl")
-include("bytes.jl")
 
+using .Common
 import .Parser.parse
 
 # These are temporary ways to bypass excess memory allocation
@@ -63,21 +64,6 @@ lower(x::Real) = Float64(x)
 
 const INDENT=true
 const NOINDENT=false
-const REVERSE_ESCAPES = Dict(map(reverse, ESCAPES))
-const escaped = Array(Vector{UInt8}, 256)
-for c in 0x00:0xFF
-    escaped[c + 1] = if c == SOLIDUS
-        [SOLIDUS]  # don't escape this one
-    elseif c ≥ 0x80
-        [c]  # UTF-8 character copied verbatim
-    elseif haskey(REVERSE_ESCAPES, c)
-        [BACKSLASH, REVERSE_ESCAPES[c]]
-    elseif iscntrl(Char(c)) || !isprint(Char(c))
-        UInt8[BACKSLASH, LATIN_U, hex(c, 4)...]
-    else
-        [c]
-    end
-end
 
 type State{I}
     indentstep::Int
@@ -134,14 +120,14 @@ end
 
 function print_escaped(io::IO, s::AbstractString)
     @inbounds for c in s
-        c <= '\x7f' ? Base.write(io, escaped[UInt8(c) + 0x01]) :
+        c <= '\x7f' ? Base.write(io, ESCAPED_ARRAY[@compat UInt8(c) + 0x01]) :
                       Base.print(io, c) #JSON is UTF8 encoded
     end
 end
 
 function print_escaped(io::IO, s::Compat.UTF8String)
     @inbounds for c in s.data
-        Base.write(io, escaped[c + 0x01])
+        Base.write(io, ESCAPED_ARRAY[c + 0x01])
     end
 end
